+++ apps/dashboard/src/components/ui/HeaderUserMenu.jsx
@@ -4,7 +4,7 @@ import useToggle from "hooks/useToggle";
 import { useContext } from "react";
 import { useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
-import KeyCloakService from "services/keycloak";
+import AuthService from "services/auth";
 import ExternalLink from "./ExternalLink";
 
 const HeaderUserMenu = () => {
@@ -16,9 +16,17 @@ const HeaderUserMenu = () => {
     return (
         <div className={`dropdown ${value ? "is-active" : ""}`}>
             <div className="dropdown-trigger">
-                <button className="button" aria-haspopup="true" aria-controls="dropdown-menu" onClick={toggle}>
+                <button
+                    className="button"
+                    aria-haspopup="true"
+                    aria-controls="dropdown-menu"
+                    onClick={toggle}
+                >
                     <span>
-                        <UserAvatar email={user.email} /> <span color="gray.500">{KeyCloakService.getUsername()}</span>{" "}
+                        <UserAvatar email={user.email} />{" "}
+                        <span color="gray.500">
+                            {AuthService.getUsername()}
+                        </span>{" "}
                         (<span>{user.role}</span>)
                     </span>
                     <span className="icon is-small">
@@ -29,7 +37,9 @@ const HeaderUserMenu = () => {
             <div className="dropdown-menu" id="dropdown-menu" role="menu">
                 <div className="dropdown-content">
                     <div className="dropdown-item">
-                        <ExternalLink href={KeyCloakService.getProfileUrl()}>Identity settings</ExternalLink>
+                        <ExternalLink href={AuthService.getProfileUrl()}>
+                            Identity settings
+                        </ExternalLink>
                     </div>
                     <Link className="dropdown-item" to={`/users/${user.id}`}>
                         <div>Your profile</div>